Move from SQLModel to raw SQLAlchemy for simplicity's sake
(we may revisit SQLModel at a later stage with FastAPI)

Rework the locations of models and routes to make a bit more sense


#20
