+++ instance/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitkeep
