+++ slurp/__init__.py
@@ -1,171 +1,20 @@
 import ast
 import os
-import tempfile
 import tomllib
-from typing import Generator
 
-from flask import (
-    Blueprint,
-    Flask,
-    current_app,
-    render_template,
-    request,
-    stream_template,
-)
-from flask_wtf import FlaskForm
-from sqlmodel import Session, select
-from wtforms import SelectField, StringField, URLField
-from wtforms.validators import URL, AnyOf, DataRequired
+from flask import Flask
 
 from slurp.api import api_blueprint
-from slurp.db import create_db_and_tables, engine
+from slurp.db import create_db_and_tables
 from slurp.fetchers import (
     fetchers,
-    fetchers_for_url,
 )
 from slurp.fetchers.cobalt import CobaltFetcher
 from slurp.fetchers.exceptions import FetcherMisconfiguredError
 from slurp.fetchers.get_iplayer import BBCiPlayerFetcher
-from slurp.fetchers.types import (
-    FetcherMediaAvailable,
-    FetcherMediaMetadataAvailable,
-    FetcherProgressReport,
-    Format,
-)
 from slurp.fetchers.ytdlp import YTDLPFetcher
-from slurp.finaliser import finalise
 from slurp.helpers import format_duration
-from slurp.models.task import FetchTask
-
-
-class DownloadForm(FlaskForm):
-    url = URLField("url", validators=[DataRequired(), URL()])
-    slug = StringField("slug", validators=[DataRequired()])
-    format = SelectField(
-        "format",
-        choices=[(v.name, v.value) for v in Format],
-        validators=[DataRequired(), AnyOf([v.name for v in Format])],
-    )
-    directory = SelectField("directory", validators=[DataRequired()])
-
-
-main_blueprint = Blueprint("main", __name__, template_folder="templates")
-
-
-@main_blueprint.get("/")
-def index():
-    form = DownloadForm(request.args)
-    form.directory.choices = current_app.config["OUTPUTS"]
-    with Session(engine) as session:
-        allTasks = session.exec(select(FetchTask)).all()
-
-        return render_template(
-            "index.html", form=form, fetchers=fetchers, allTasks=allTasks
-        )
-
-
-def stream_fetch(url: str, format: Format, target: str, slug: str) -> Generator[str]:
-    """stream_fetch is a rudimentary function that fetches the given media, yielding to generated HTML strings for progress reports."""
-    fetchers = fetchers_for_url(url)
-    if len(fetchers) == 0:
-        yield "<article class='fetcher-outcome fetcher-progress-message-level-error'>☹️ Slurp failed - no fetchers can handle this request. Check URL?</article>"
-        return
-
-    # Work in a temporary directory that gets torn down at the completion of this slurp run
-    with tempfile.TemporaryDirectory(
-        dir=current_app.config.get("OUTPUT_TEMP", None)
-    ) as tmp_dir:
-        success: bool = False
-        media_path: str | None = None
-        for idx, fetcher in enumerate(fetchers):
-            yield f"<code class='fetcher-progress-message'>🛫 {'Trying Fetch again' if idx > 0 else 'Fetching'} with {fetcher.name}...</code>"
-            for event in fetcher.fetch(url, format, tmp_dir, slug):
-                match event:
-                    case FetcherMediaMetadataAvailable() as e:
-                        # Metadata for this fetch now available.
-                        yield render_template(
-                            "elements/media.html", metadata=e.metadata
-                        )
-                    case FetcherMediaAvailable() as e:
-                        media_path = e.path
-                    case FetcherProgressReport() as e:
-                        yield render_template("elements/progress_report.html", event=e)
-                        if e.typ == "finish":
-                            if e.status == 0:
-                                # Success
-                                success = True
-                            else:
-                                yield f"<code><b>🛬 Fetcher failed! Reason: {e.message}</b></code>"
-
-            if success:
-                yield "<article class='fetcher-outcome fetcher-progress-message-level-success'>✅ Media fetch successful</article>"
-                break
-        if not success:
-            yield "<article class='fetcher-outcome fetcher-progress-message-level-error'>☹️ Slurp failed - out of available fetchers.</article>"
-
-        # Run the finalisers.
-        yield "<article class='fetcher-outcome fetcher-progress-message-level-info'>🚚 Finalising slurp...</article>"
-        final_path: str | None = None
-        try:
-            for event in finalise(media_path, target):
-                match event:
-                    case FetcherProgressReport() as e:
-                        yield render_template("elements/progress_report.html", event=e)
-                    case FetcherMediaAvailable() as e:
-                        final_path = e.path
-        except Exception as e:
-            yield f"<article class='fetcher-outcome fetcher-progress-message-level-error'>💣 Failed to finalise media: {e}</article>"
-            return
-        yield f"<article class='fetcher-outcome fetcher-progress-message-level-success'>🥤 Media slurped to {final_path}</article>"
-        return
-
-
-def fetch_job_create(url: str, format: Format, target: str, slug: str) -> FetchTask:
-    task = FetchTask(url=url, format=format, target=target, slug=slug)
-    with Session(engine) as session:
-        session.add(task)
-        session.commit()
-
-    return task
-
-
-def fetch_work(task: FetchTask):
-    # This is the beginnings of what will be the task runner thread.
-    # Right now it just wraps the HTML generator in some database updates - but eventually
-    # this will simply take the task ID, pull it from DB, then trust those details
-    # to complete the fetch - putting assoc. events onto the event bus.
-    with Session(engine) as session:
-        task.status = FetchTask.TaskStatus.running
-        session.add(task)
-        session.commit()
-
-        yield stream_fetch(
-            url=task.url, format=task.format, target=task.target, slug=task.slug
-        )
-
-        task.status = FetchTask.TaskStatus.success
-        session.add(task)
-        session.commit()
-
-        return
-
-
-@main_blueprint.post("/download")
-def download():
-    form = DownloadForm()
-    form.directory.choices = current_app.config["OUTPUTS"]
-    if not form.validate_on_submit():
-        return form.errors
-    task = fetch_job_create(
-        form.url.data,
-        Format[form.format.data],
-        form.directory.data,
-        form.slug.data,
-    )
-    return stream_template(
-        "download.html",
-        output=fetch_work(task),
-    )
+from slurp.routes import main_blueprint
 
 
 def create_app(config_filename: str = "config.toml") -> Flask:
@@ -196,8 +45,11 @@ def create_app(config_filename: str = "config.toml") -> Flask:
             "SECRET_KEY has not been set - THIS IS NOT SECURE! Are you providing valid configuration?"
         )
 
+    # Set appropriate SQLAlchemy connection URI.
+    app.config["SQLALCHEMY_DATABASE_URI"] = f"sqlite:///{app.config['DB_LOCATION']}"
+
     # Initialise database.
-    create_db_and_tables()
+    create_db_and_tables(app)
 
     if isinstance(app.config["OUTPUTS"], str):
         app.logger.error(f"Outputs pre-split: {app.config['OUTPUTS']}")
