+++ slurp/api/fetchtasks.py
@@ -1,35 +1,7 @@
-# from flask.views import MethodView
-# from flask_smorest import Blueprint
 from flask_restx import Namespace, Resource, fields
-from sqlmodel import Session, select
 
-from slurp.db import engine
+from slurp.db import db
 from slurp.models.task import FetchTask
-
-# tasks_blueprint = Blueprint(
-#     "tasks", "tasks", url_prefix="/tasks", description="Operations on Tasks"
-# )
-
-
-# @tasks_blueprint.route("/")
-# class Tasks(MethodView):
-#     @tasks_blueprint.response(200, FetchTask)
-#     def get(self):
-#         with Session(engine) as session:
-#             statement = select(FetchTask)
-#             if id is not None:
-#                 return session.exec(statement.where(FetchTask.id == id)).first()
-#             else:
-#                 return session.exec(statement).all()
-#
-#
-# @tasks_blueprint.route("/<task_id>")
-# class TasksById(MethodView):
-#     @tasks_blueprint.response(200, FetchTask)
-#     def get(self, task_id):
-#         """Get Task by ID"""
-#         with Session(engine) as session:
-#             return session.get(FetchTask, task_id)
 
 api = Namespace("tasks", description="Fetch tasks")
 
@@ -54,5 +26,4 @@ class TaskList(Resource):
     @api.doc("list_tasks")
     @api.marshal_list_with(fetchTask)
     def get(self):
-        with Session(engine) as session:
-            return session.exec(select(FetchTask)).all()
+        db.session.execute(db.select(FetchTask)).fetchall()
