+++ slurp/models/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["FetchTask"]
+
+from .task import FetchTask
