+++ slurp/models/base.py
@@ -1,26 +1,28 @@
 import json
 from datetime import datetime, timezone
 
-from sqlalchemy import TIMESTAMP
-from sqlmodel import Field, SQLModel
+from sqlalchemy import func
+from sqlalchemy.orm import Mapped, mapped_column
+
+from slurp.db import db
 
 
-class BaseModel(SQLModel):
+class BaseModel(db.Model):
     """
     Base model
     """
 
     __abstract__ = True
     # Ordinarily I prefer UUIDs for all objects, but as we want SQLite we'll keep it simple and go auto-increment.
-    id: int | None = Field(default=None, primary_key=True)
-    ts_created: datetime = Field(default=datetime.now(timezone.utc), nullable=False)
-    ts_updated: datetime | None = Field(
-        default_factory=lambda: datetime.now(timezone.utc),
+    id: Mapped[int] = mapped_column(primary_key=True)
+    ts_created: Mapped[datetime] = mapped_column(
+        default=datetime.now(timezone.utc), server_default=func.now(), nullable=False
+    )
+    ts_updated: Mapped[datetime] = mapped_column(
+        default=datetime.now(timezone.utc),
+        server_default=func.now(),
+        onupdate=func.current_timestamp(),
         nullable=False,
-        sa_column_kwargs={
-            "onupdate": lambda: datetime.now(timezone.utc),
-        },
-        sa_type=TIMESTAMP(timezone=True),
     )
 
     @classmethod
@@ -50,14 +52,3 @@ class BaseModel(SQLModel):
         Get the entity that matches the id
         """
         return cls.query.get(_id)
-
-    # This must be overridden by derived classes
-    def json(self) -> json:
-        """
-        Get model data in JSON format
-        """
-        return {
-            "id": self.id,
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-        }
