+++ slurp/models/task.py
@@ -1,36 +1,41 @@
 import enum
 
-from sqlalchemy import text
-from sqlmodel import Field
+from sqlalchemy.orm import Mapped, mapped_column
 
 from slurp.fetchers.types import Format
 from slurp.models.base import BaseModel
 
 
-class FetchTask(BaseModel, table=True):
+class FetchTask(BaseModel):
     # URL is the fully qualified link to the media to be fetched.
-    url: str = Field(index=True)
+    url: Mapped[str] = mapped_column(index=True)
     # Slug is the desired output media name.
-    slug: str = Field(index=True)
+    slug: Mapped[str] = mapped_column(index=True)
+    # Target is the desired destination identifier. If not set, uses the app-configured default.
+    target: Mapped[str | None] = mapped_column()
 
     # Format is the desired format to fetch the URL in.
-    format: Format = Field(
-        sa_column_kwargs={"server_default": text(f"'{Format.VIDEO_AUDIO.value}'")},
+    format: Mapped[Format] = mapped_column(
         default=Format.VIDEO_AUDIO.value,
         nullable=False,
     )
-    # Target is the desired destination identifier. If not set, uses the app-configured default.
-    target: str | None
 
     class TaskStatus(str, enum.Enum):
+        # "created" tasks are awaiting processing or assignment to a worker.
         created = "created"
+        # "Running" tasks are in the process of being fetched.
         running = "running"
+        # "Success" tasks completed their fetch successfully.
         success = "success"
+        # "Failed" tasks failed to process for some reason.
         failed = "failed"
+        # "Completed" tasks is one where it finished executing, but we don't know the outcome for some reason.
+        completed = "completed"
+        # "Unknown" tasks are ones where the execution state is a mystery to us.
+        unknown = "unknown"
 
     # Status
-    status: TaskStatus = Field(
-        sa_column_kwargs={"server_default": text(f"'{TaskStatus.created.value}'")},
+    status: Mapped[TaskStatus] = mapped_column(
         default=TaskStatus.created.value,
         nullable=False,
     )
